feat: 챌린지 상세를 탭 뷰(개요·일지·통계·참여자)로 분할 (#170)
히어로(썸네일·제목·기간·참여 CTA)는 탭 위에 항상 고정 노출하고, 세로로
길던 카드 스택을 4개 탭으로 재구성한다.

- 개요: 챌린지 소개 + 목표(규칙) 카드
- 일지: 무한 스크롤 일지 리스트 + 날짜 필터(통계 히트맵에서 진입)
- 통계: 참여율·완료 목표수·일자별 추이·히트맵 캘린더
- 참여자: 랭킹(스트릭/완료 목표수)·찌르기 + 전체 목록 이동

탭 상태는 URL ?tab= 으로 보존(기본 개요)하고, DS Tabs 로 렌더한다.
각 탭 콘텐츠는 활성 시에만 마운트돼 비활성 탭의 일지/통계 쿼리는
실행되지 않는다(상세 진입 시 요청 폭증 방지).

- 일지 리스트 본문을 ChallengeDiaryList 로 추출해 상세 탭·standalone
  라우트가 공유
- 우측 rail 은 진행률 CTA + 나가기만 유지(리더보드는 참여자 탭으로 이동)
- 상세 page 를 Suspense 로 감싸 useSearchParams CSR bailout 처리
- 스켈레톤에 탭 바 반영, 사용하지 않는 ChallengeDiaryGrid /
  useChallengeDiaryList / diaries 쿼리키 제거
