+++ src/app.feature/challenge/board/consts/queryKeys.ts
@@ -21,15 +21,6 @@ export const CHALLENGE_QUERY_KEYS = {
   checkWrite: (challengeId: number) =>
     [...CHALLENGE_QUERY_KEYS.all, 'check-write', challengeId] as const,
   challengeDiaries: () => [...CHALLENGE_QUERY_KEYS.all, 'diaries'] as const,
-  diaries: (
-    challengeId: number,
-    params?: { page?: number; size?: number; date?: string }
-  ) =>
-    [
-      ...CHALLENGE_QUERY_KEYS.challengeDiaries(),
-      challengeId,
-      params,
-    ] as const,
   diariesInfinite: (
     challengeId: number,
     params?: { size?: number; date?: string }
